backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File Upload and Processing
python-multipart==0.0.6
aiofiles==23.2.1

# Excel Processing
openpyxl==3.1.2
pandas==2.1.3
xlrd==2.0.1

# Data Processing and Validation
numpy==1.25.2
pydantic==2.5.0

# Environment and Configuration
python-dotenv==1.0.0



# Logging and Utilities
loguru==0.7.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: Progress tracking for large files
tqdm==4.66.1